add fetchingPaymentInfo state
